Add files via upload
Here are the files for my final project for Unsupervised Algorithms in Machine Learning. The final project is about creating models to properly classify types of dry beans. There is a Jupyter notebook file that contains my project. There is an Excel file that contains the dry bean data set. There is a text file that contains information about the dry bean data set. The link and citation for the dry bean data set is contained in my Jupyter notebook.